Minor refactors and add docs
